add from_document
