+++ src/lib.rs
@@ -196,7 +196,7 @@ pub use self::{
         Regex,
         Timestamp,
     },
-    de::{from_bson, Deserializer},
+    de::{from_bson, from_document, Deserializer},
     decimal128::Decimal128,
     ser::{to_bson, to_document, Serializer},
 };
